Implement Rest API, JPA layer, and expand tests
This commit provides a significant extension, introducing the Rest API layer that can handle GET requests for candidates by state and district. The persistence layer was created using JPA, with a specification of repo and DTO for candidates. Moreover, the test environment was extended with saving data, reading and test API endpoint functions.